# openapi.yml
...
openapi: 3.0.0
paths:
...
  /media-jobs/webhook/{jobId}/{token}:
    post:
      operationId: MediaJobsWebhookController_handle
      parameters:
        - name: jobId
          required: true
          in: path
          schema:
            type: string
        - name: token
          required: true
          in: path
          schema:
            type: string
      responses:
        "201":
          description: ""
      tags:
        - Media Jobs
  /webhooks/comms/platform/{identifier}:
    post:
      operationId: CommsWebhooksController_handlePlatform
      parameters:
        - name: identifier
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: object
      tags:
        - Comms Webhooks
  /webhooks/comms/{identifier}/{token}:
    post:
      operationId: CommsWebhooksController_handle
      parameters:
        - name: identifier
          required: true
          in: path
          schema:
            type: string
        - name: token
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: object
      tags:
        - Comms Webhooks
...
info:
  title: ValidPost API
  description: OpenAPI description of the ValidPost backend (NestJS). Generated from the controllers — do not hand-edit; run `pnpm run openapi:generate`.
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    api-key:
      type: apiKey
      in: header
      name: Authorization
    bearer:
      scheme: bearer
      bearerFormat: Token
      type: http
...